+++ db/updates/update_5.9-6.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `suppliers` CHANGE `stock` `stock` DECIMAL(10,3) NOT NULL;