+++ func/getIngSupplier.php
@@ -22,7 +22,7 @@ function getPrefSupplier($ingID,$conn){
 	$ing = mysqli_fetch_array(mysqli_query($conn, "SELECT price,ingSupplierID,size,supplierLink FROM suppliers WHERE ingID = '$ingID' AND preferred = '1'"));
 	$sup = mysqli_fetch_array(mysqli_query($conn, "SELECT name FROM ingSuppliers WHERE id = '".$ing['ingSupplierID']."'"));
 		
-	$result = array_merge($ing, $sup);
+	$result = array_merge((array)$ing, (array)$sup);
 
 	return $result;
 }