+++ pages/viewSummary.php
@@ -145,3 +145,103 @@ html {
 
 <p>&nbsp;</p>
 
+<!--Configure View-->
+
+<div class="modal fade" id="conf_view" tabindex="-1" role="dialog" aria-labelledby="conf_view" aria-hidden="true">
+  <div class="modal-dialog modal-conf-view" role="document">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h5 class="modal-title" id="conf_view">Choose which notes will be displayed</h5>
+        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+          <span aria-hidden="true">&times;</span>
+        </button>
+      </div>
+      <div class="modal-body">
+   	    <div id="confViewMsg"></div>
+          <form action="javascript:update_view()" id="form1">   
+           <div class="conf_tbl">
+            <table width="100%" border="0">
+              <tr>
+                <td colspan="2"><strong>Top notes</strong><hr /></td>
+              </tr>
+              <?php foreach ($top_cat as $x){
+						if (!is_numeric(array_search($x['name'],$top_ex ))){
+				?>
+              <tr>
+				<td width="54%" height="29" ex_top_ing_name="<?=$x['name']?>"><?=$x['name']?></td>
+                <td width="46%"><input name="ex_top_ing" class="ex_ing" type="checkbox" id="<?=str_replace(' ', '_',$x['ing'])?>" value="<?=str_replace(' ', '_',$x['ing'])?>" checked="checked" /></td>
+              </tr>
+              <?php }else{ ?>
+			  <tr>
+				<td width="54%" ex_top_ing_name="<?=$x['name']?>"><?=$x['name']?></td>
+                <td width="46%"><input name="ex_top_ing" class="ex_ing" type="checkbox" id="<?=str_replace(' ', '_',$x['ing'])?>" value="<?=str_replace(' ', '_',$x['ing'])?>" /></td>
+              </tr>
+			 <?php 
+			 	}
+			  }
+			  ?>
+             </table>
+            </div>
+             
+           <div class="conf_tbl">
+            <table width="100%" border="0">
+              <tr>
+                <td colspan="2"><p><strong>Heart notes</strong></p><hr /></td>
+              </tr>
+              <?php foreach ($heart_cat as $x){
+						if (!is_numeric(array_search($x['name'],$heart_ex ))){
+			   ?>
+              <tr>
+				<td width="40%" height="29"><?=$x['name']?></td>
+                <td width="51%"><input name="ex_heart_ing" class="ex_ing" type="checkbox" id="<?=str_replace(' ', '_',$x['ing'])?>" value="<?=str_replace(' ', '_',$x['ing'])?>" checked="checked" /></td>
+              </tr>
+              <?php }else{ ?>
+              <tr>
+				<td><?=$x['name']?></td>
+                <td width="51%"><input name="ex_heart_ing" class="ex_ing" type="checkbox" id="<?=str_replace(' ', '_',$x['ing'])?>" value="<?=str_replace(' ', '_',$x['ing'])?>" /></td>
+              </tr>
+              <?php 
+			 	}
+			  }
+			  ?>
+             </table> 
+            </div>
+
+            <div class="conf_tbl">
+             <table width="100%" border="0">
+              <tr>
+                <td colspan="2"><p><strong>Base notes</strong></p><hr /></td>
+              </tr>
+              <?php foreach ($base_cat as $x){
+						if (!is_numeric(array_search($x['name'],$base_ex ))){
+			  ?>
+              <tr>
+				<td width="40%" height="29"><?=$x['name']?></td>
+                <td width="60%"><input name="ex_base_ing" class="ex_ing" type="checkbox" id="<?=str_replace(' ', '_',$x['ing'])?>" value="<?=str_replace(' ', '_',$x['ing'])?>" checked="checked" /></td>
+              </tr>
+             <?php }else{ ?>
+              <tr>
+				<td><?=$x['name']?></td>
+                <td width="60%"><input name="ex_base_ing" class="ex_ing" type="checkbox" id="<?=str_replace(' ', '_',$x['ing'])?>" value="<?=str_replace(' ', '_',$x['ing'])?>" /></td>
+              </tr>
+              <?php 
+			 	}
+			  }
+			  ?>
+            </table>
+            </div>
+            <table width="100%" border="0">
+              <tr>
+                <td>            
+  					<div class="modal-footer">
+     	  				<button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
+	 	  				<input type="submit" name="button" class="btn btn-primary" id="btnUpdate" value="Save">
+   		  			</div>
+           		 </td>
+    			</tr>
+			</table>
+          </form>
+    </div>
+  </div>
+  </div>
+</div>