+++ releasenotes.md
@@ -1,2 +1,3 @@
 Whats New:
 ----------------------
+- Better stock management when making a formula